profile_pruner follows writing format, and query before pruning+fine_tuning

+++ sources/TorchPruner/profile_pruner.py
@@ -1,18 +1,20 @@
 import os
 
-from prune_resnet import prune_resnet_and_fine_tune
+from prune_resnet import prune_resnet_and_fine_tune, query_profiled_result
 from RecordIO.RecordIO import save_to_file
 from ResNet.variables import ROOT_PATH
 from ResNet.Hyperparameters import Hyperparameters, get_hp_formal_cifar10, get_hp_test_cifar10, get_hp_test_cifar10_fast
 
 from torchvision.models import resnet18
+from RecordIO.WritingInfo import WritingInfo, get_output_file_name
 
 
-def create_csv_file(file_name):
+def get_csv_file(file_name):
     file_path = os.path.join(ROOT_PATH, "profile_data",
-                             file_name+get_date()+".csv")
-    with open(file_path, 'w') as file:
-        file.write("Pruning_ratio, Accuracy, Latency\n")
+                             file_name)
+    if not os.path.exists(os.path.dirname(file_path)):
+        with open(file_path, 'w') as file:
+            file.write("Pruning_ratio, Accuracy, Latency\n")
     return file_path
 
 
@@ -20,19 +22,32 @@ def get_date():
     import datetime
     return datetime.datetime.now().strftime("%Y-%m-%d-%H-%M-%S")
 
+def find_record(model_name, training_data_noise, pruning_ratio):
+    file_name = get_output_file_name(model_name, training_data_noise)
+    file_path = get_csv_file(file_name)
+    if not os.path.exists(file_path):
+        return None, None
+    
+    return query_profiled_result(pruning_ratio, file_name)
 
-def profile_with_pruning_ratio(model, hyperparameters, record_file_name, granularity=0.1):
-    """Profile the model with pruning ratio, and save it to record_file_name"""
 
-    file_path = create_csv_file(record_file_name)
+def profile_with_pruning_ratio(model, hyperparameters_model, model_name, training_data_noise,\
+                                granularity=0.1):
+    """
+    Profile the model with different pruning ratioes, and save it to profile_data folder
+    """
+    record_file_name = get_output_file_name(model_name=model_name, training_data_noise=training_data_noise)
+    file_path = get_csv_file(record_file_name)
     for pruning_ratio in range(0, 100, int(granularity*100)):
         pruning_ratio = pruning_ratio / 100.0
-        accuracy, latency = prune_resnet_and_fine_tune(
-            model, pruning_ratio, hyperparameters)
+        accuracy, latency = find_record(model_name, training_data_noise, pruning_ratio)
+        if not accuracy and not latency:
+            accuracy, latency = prune_resnet_and_fine_tune(
+                model, pruning_ratio, hyperparameters_model)
+            save_to_file([pruning_ratio], [accuracy, latency], file_path)
         print(
             f"Pruning ratio: {pruning_ratio}, accuracy: {accuracy}, latency: {latency}")
-        save_to_file([pruning_ratio], [accuracy, latency], file_path)
-
+    
 
 if __name__ == "__main__":
     # hyperparameters = get_hp_test_cifar10_fast()
@@ -41,4 +56,4 @@ if __name__ == "__main__":
     model = resnet18(weights='ResNet18_Weights.DEFAULT')
     # model = resnet18(pretrained=True)
     profile_with_pruning_ratio(model, hyperparameters,
-                               "profile_test_fast_resnet18_pruning_ratio_0.1_noise")
+                               "resnet18", "0.0")
